move metrics listener to stages crate
