--- crates/metrics/src/listener/events/mod.rs
@@ -1,5 +1,0 @@
-//! Metric events and supporting types.
-
-mod sync;
-
-pub use sync::*;
