+++ crates/stages/src/metrics/mod.rs
@@ -1,0 +1,5 @@
+mod listener;
+mod sync_metrics;
+
+pub use listener::{MetricEvent, MetricEventsSender, MetricsListener};
+use sync_metrics::*;
